DataBlockHashIndex: Add use_data_block_hash_index flag to Block::NewIterator()
Summary:

This is to distinguish a point lookup DataBlockIter and a seek
DataBlockIter. Only a point lookup DataBlockIter can use
DataBlockHashIndex.

Besides, IndexBlockIter should never use DataBlockHashIndex.

The following three statements are equivalent:

A DataBlockIter is a point lookup DataBlockIter
iff. use_data_block_hash_index is set to true in Block::NewIterator()
iff. DataBlockIter is init'ed with data_block_hash_index is not nullptr

Point lookup DataBlockIter will not set restart_index_ to the
right place. So Prev() cannot be used in a point lookup DataBlockIter.
It is the callers' responsibility not using Prev() in a point lookup
DataBlockIter.