DataBlockHashIndex: Space Optimization Part I, Remove TAG
Summary:
Space Optimization;

we can store the restart index in the hash table. We do not allow
collision. If there is a collision, we fall back to binary seek.
So there is only one entry per bucket. We can test out different
utilization ratio of the hash table.

Several possible outcome of the HashTable Query: a) only one entry,
goes to the restart interval; b) no entry, return not found; and
3) multiple entry, hash table report collision, and tell the caller
to fall back to binary seek.  Taking the assumptions that 1Byte is
used to store the restart interval index, we reserve two index as
special flag: kNoEntry = 255 and kCollision = 254. Otherwise, the
entry is valid. This is less expensive to borrow the most
significant 2bits to indicate the entry type.

Part I: Remove TAG
Part II: Collision marker and fall back
Part III: Absort the restart index into hash index table