DataBlockHashIndex: avoiding expensive iiter->Next when handling hash kNoEntry
Summary:
Summary:
When returning kNoEntry from HashIndex lookup, previously we invalidate the
biter by set current_=restarts_, so that the search can continue to the next
block in case the search result may reside in the next block.

There is one problem: when we are searching for a missing key, if the search
finds a kNoEntry and continue the search to the next block, there is also a
non-trivial possibility that the HashIndex return kNoEntry too, and the
expensive index iterator Next() will happen several times for nothing.

The solution is that if we find a kNoEntry, we pretend the key is in the last
restart_interval. The linear search for a matching user key will handle it
correctly.